requirements/requirements_test.txt
coverage
coveralls
factory_boy==2.3.1
mock
pytest
pytest-django
flake8
selenium
tox
six
lxml
cssselect
